Before: nil.blank?    => true nil._?.blank? => false
because calling blank? on a SafeNil doesn't get to method_missing,
because Facets defines blank? in Kernel so it's defined for every
object, and is true because !!SafeNil.instance is true.

But that doesn't make any intuitive sense, so we define SafeNil#blank?
such that
nil.blank?    => true
nil._?.blank? => true